Fixed API token Generation

+++ api/views.py
@@ -1,9 +1,9 @@
 from index.models import Record
-from .serializer import RecordSerializer
+from .serializer import RecordSerializer, TokenSerializer
 from django_filters.rest_framework import DjangoFilterBackend
 from rest_framework import generics, filters
 from rest_framework.authentication import TokenAuthentication
-from rest_framework.permissions import IsAuthenticatedOrReadOnly, IsAuthenticated
+from rest_framework.permissions import IsAuthenticatedOrReadOnly, IsAuthenticated, AllowAny
 from knox.auth import TokenAuthentication
 from .filters import RecordFilter
 
@@ -26,15 +26,23 @@ class RecordListCreateView(generics.ListCreateAPIView):
 
 class TokenView(generics.GenericAPIView):
     authentication_classes = [TokenAuthentication]
-    permission_classes = [IsAuthenticated]
+    permission_classes = [AllowAny]
+    serializer_class = TokenSerializer
 
-    def post(self, request, *args, **kwargs):
-        serializer = TokenSerializer(data=request.data)
+    def post(self, request, format=None):
+        serializer = self.get_serializer(data=request.data)
         serializer.is_valid(raise_exception=True)
         user = serializer.validated_data['user']
-        token, _ = AuthToken.objects.create(user)
-        return Response({'token': token}, status=status.HTTP_201_CREATED)
+        token = AuthToken.objects.create(user)
+        return Response({
+            'token': token,
+        })
 
     def delete(self, request, *args, **kwargs):
-        request.user.auth_token.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
+        user_token = request.auth
+        if user_token:
+            request.user.auth_token.delete()
+            user_token.delete()
+            return Response({'detail': 'Token invalidated'})
+        else:
+            return Response({'detail': 'No token provided'})
